requirements.txt
Flask
bcrypt
mysql-connector-python